+++ src/main/java/cl/cavallinux/jisocreator/action/main/SaveAsIsoAction.java
@@ -5,7 +5,6 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
-import java.util.Optional;
 
 import org.apache.commons.lang3.StringUtils;
 import org.eclipse.jface.action.Action;
@@ -76,10 +75,7 @@ public class SaveAsIsoAction extends Action implements IFileManagementAction {
 
     private IsoFileSystem obtainIsoFileSystem() {
         if (commandLineMode) {
-            Optional<IsoFileSystem> deserializedISOFileSystem = IOManager.INSTANCE.getIsoFilesystemParser()
-                    .deserialize(inputXMLLayoutFile);
-            log.info("Deserialized isofilesystem: {}", deserializedISOFileSystem);
-            return (IsoFileSystem) IOManager.INSTANCE.getIoUtils().parseXMLFileToObject(inputXMLLayoutFile);
+            return (IsoFileSystem) IOManager.INSTANCE.getIsoFilesystemParser().deserialize(inputXMLLayoutFile).get();
         } else {
             return (IsoFileSystem) GUIManager.INSTANCE.getMainWindow().getIsoExplorer().getIsoDirectoriesTree()
                     .getInput();
