+++ src/main/java/cl/cavallinux/jisocreator/model/isoexplorer/impl/TreeNode.java
@@ -3,11 +3,6 @@ package cl.cavallinux.jisocreator.model.isoexplorer.impl;
 import java.util.ArrayList;
 import java.util.List;
 
-import com.fasterxml.jackson.annotation.JsonBackReference;
-import com.fasterxml.jackson.annotation.JsonManagedReference;
-import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlElementWrapper;
-import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty;
-
 import cl.cavallinux.jisocreator.model.isoexplorer.decl.ITreeNode;
 
 /**
@@ -19,11 +14,7 @@ import cl.cavallinux.jisocreator.model.isoexplorer.decl.ITreeNode;
  * @version 0.0.2
  */
 public abstract class TreeNode implements ITreeNode {
-    @JsonBackReference
     protected ITreeNode parent;
-    @JsonManagedReference
-    @JacksonXmlElementWrapper(localName = "children")
-    @JacksonXmlProperty(localName = "entry")
     protected List<ITreeNode> children;
 
     protected TreeNode(ITreeNode parent) {
@@ -39,7 +30,7 @@ public abstract class TreeNode implements ITreeNode {
     public Object[] getChildren() {
         return children.toArray();
     }
-
+    
     @Override
     public ITreeNode getParent() {
         return parent;
