+++ src/main/java/cl/cavallinux/jisocreator/util/IOUtils.java
@@ -1,37 +1,22 @@
 package cl.cavallinux.jisocreator.util;
 
-import java.io.FileInputStream;
-import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.file.Files;
 import java.nio.file.Paths;
-import java.util.Objects;
 import java.util.Properties;
 import java.util.UUID;
 
-import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.Strings;
 import org.eclipse.jface.preference.PreferenceStore;
 import org.eclipse.swt.SWT;
 
-import com.thoughtworks.xstream.XStream;
-import com.thoughtworks.xstream.XStreamException;
-import com.thoughtworks.xstream.converters.reflection.PureJavaReflectionProvider;
-import com.thoughtworks.xstream.security.NoTypePermission;
-import com.thoughtworks.xstream.security.NullPermission;
-import com.thoughtworks.xstream.security.PrimitiveTypePermission;
-
-import cl.cavallinux.jisocreator.model.isoexplorer.impl.IsoFileSystem;
-import cl.cavallinux.jisocreator.model.isoexplorer.impl.IsoTreeNode;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 public class IOUtils {
     private PreferenceStore store;
     private Properties defaultProperties;
-    @Deprecated(since = "0.1.6", forRemoval = true)
-    private XStream xStreamParser;
     private final static String JISOCREATOR_CONFIG_DIR;
     private final static String JISOCREATOR_DEFAULTCONFIG_FILENAME;
     private final static String JISOCREATOR_CONFIG_FILENAME;
@@ -52,10 +37,6 @@ public class IOUtils {
         MKISOFS_ISOFILESYSTEM_APPLICATIONID = String.format("%s", IOUtils.class.getPackage().getImplementationTitle());
     }
 
-    public IOUtils() {
-        loadXMLParser();
-    }
-
     public String generateInitialVolumeID() {
         String isoFileSystemVolumeID = UUID.randomUUID().toString();
         isoFileSystemVolumeID = isoFileSystemVolumeID.replace("-", "");
@@ -64,29 +45,6 @@ public class IOUtils {
 
     public String generateIsoFilesystemApplicationID() {
         return MKISOFS_ISOFILESYSTEM_APPLICATIONID;
-    }
-
-    @Deprecated(since = "0.1.6", forRemoval = true)
-    public Object parseXMLFileToObject(String path) {
-        try (FileInputStream fis = new FileInputStream(path)) {
-            IsoFileSystem iso = (IsoFileSystem) xStreamParser.fromXML(fis);
-            repairApplicationIDAndPublisherID(iso);
-            return iso;
-        } catch (IOException | XStreamException e) {
-            log.error("Error parsing XML", e);
-            return null;
-        }
-    }
-
-    @Deprecated(since = "0.1.6", forRemoval = true)
-    public boolean saveObjectToXML(Object objectToParse, String path) {
-        try (FileOutputStream fos = new FileOutputStream(path)) {
-            xStreamParser.toXML(objectToParse, fos);
-            return true;
-        } catch (IOException e) {
-            log.error("Error saving XML", e);
-            return false;
-        }
     }
 
     public PreferenceStore getStore() {
@@ -108,33 +66,6 @@ public class IOUtils {
             log.warn("Loading defaults", e);
             loadPreferencesFromBackup();
         }
-    }
-
-    @Deprecated(since = "0.1.6", forRemoval = true)
-    private void loadXMLParser() {
-        xStreamParser = new XStream(new PureJavaReflectionProvider());
-        xStreamParser.addPermission(NoTypePermission.NONE);
-        xStreamParser.addPermission(NullPermission.NULL);
-        xStreamParser.addPermission(PrimitiveTypePermission.PRIMITIVES);
-        xStreamParser.allowTypesByWildcard(new String[] { "cl.cavallinux.jisocreator.model.isoexplorer.impl.**" });
-
-        xStreamParser.alias("iso9660", IsoFileSystem.class);
-        xStreamParser.alias("entry", IsoTreeNode.class);
-        xStreamParser.aliasAttribute(IsoFileSystem.class, "root", "RootEntry");
-        xStreamParser.aliasAttribute(IsoFileSystem.class, "volumeID", "volumeid");
-        xStreamParser.aliasAttribute(IsoFileSystem.class, "applicationID", "applicationid");
-        xStreamParser.aliasAttribute(IsoFileSystem.class, "isoLength", "isolength");
-        xStreamParser.aliasAttribute(IsoFileSystem.class, "publisherID", "publisherid");
-        xStreamParser.aliasAttribute(IsoFileSystem.class, "isoLength", "isolength");
-        xStreamParser.aliasAttribute(IsoTreeNode.class, "isRoot", "root");
-        xStreamParser.aliasAttribute(IsoTreeNode.class, "isoName", "isoname");
-        xStreamParser.useAttributeFor(IsoFileSystem.class, "volumeID");
-        xStreamParser.useAttributeFor(IsoFileSystem.class, "applicationID");
-        xStreamParser.useAttributeFor(IsoFileSystem.class, "isoLength");
-        xStreamParser.useAttributeFor(IsoFileSystem.class, "publisherID");
-        xStreamParser.useAttributeFor(IsoTreeNode.class, "file");
-        xStreamParser.useAttributeFor(IsoTreeNode.class, "isRoot");
-        xStreamParser.useAttributeFor(IsoTreeNode.class, "isoName");
     }
 
     private void loadPreferencesFromBackup() {
@@ -170,16 +101,5 @@ public class IOUtils {
                     WIN32_MKISOFS_BASE_PATH, Paths.get("").toAbsolutePath().toString()));
         }
         return mkisofsPath.toString();
-    }
-
-    private void repairApplicationIDAndPublisherID(IsoFileSystem iso) {
-        if (Objects.nonNull(iso)) {
-            if (StringUtils.isBlank(iso.getPublisherID())) {
-                iso.setPublisherID(UUID.randomUUID().toString());
-            }
-            if (!Strings.CI.equalsAny(MKISOFS_ISOFILESYSTEM_APPLICATIONID, iso.getApplicationID())) {
-                iso.setApplicationID(MKISOFS_ISOFILESYSTEM_APPLICATIONID);
-            }
-        }
     }
 }
